+++ DatadogInternal/Sources/NetworkInstrumentation/NetworkInstrumentationFeature.swift
@@ -153,11 +153,19 @@ extension NetworkInstrumentationFeature {
     ///   - task: The created task.
     ///   - additionalFirstPartyHosts: Extra hosts to consider in the interception.
     func intercept(task: URLSessionTask, additionalFirstPartyHosts: FirstPartyHosts?) {
+        // In response to https://github.com/DataDog/dd-sdk-ios/issues/1638 capture the current request object on the
+        // caller thread and freeze its attributes through `ImmutableRequest`. This is to avoid changing the request
+        // object from multiple threads:
+        guard let currentRequest = task.currentRequest else {
+            return
+        }
+        let request = ImmutableRequest(request: currentRequest)
+
         // Get the current trace context from all handlers.
         let traceContexts = handlers.compactMap { $0.traceContext() }
 
         queue.async { [weak self] in
-            guard let self = self, let request = task.currentRequest else {
+            guard let self = self else {
                 return
             }
 
@@ -187,7 +195,7 @@ extension NetworkInstrumentationFeature {
                 ))
             }
 
-            if let origin = request.value(forHTTPHeaderField: TracingHTTPHeaders.originField) {
+            if let origin = request.allHTTPHeaderFields?[TracingHTTPHeaders.originField] {
                 interception.register(origin: origin)
             }
 
@@ -259,7 +267,7 @@ extension NetworkInstrumentationFeature {
         interceptions[task] = nil
     }
 
-    private func extractTrace(firstPartyHosts: FirstPartyHosts, request: URLRequest) -> (traceID: TraceID, spanID: SpanID, parentSpanID: SpanID?)? {
+    private func extractTrace(firstPartyHosts: FirstPartyHosts, request: ImmutableRequest) -> (traceID: TraceID, spanID: SpanID, parentSpanID: SpanID?)? {
         guard let headers = request.allHTTPHeaderFields else {
             return nil
         }