+++ DatadogLogs/Sources/Feature/MessageReceivers.swift
@@ -55,7 +55,7 @@ internal struct LogMessageReceiver: FeatureMessageReceiver {
                 return false
             }
 
-            core.scope(for: LogsFeature.name)?.eventWriteContext { context, writer in
+            core.scope(for: LogsFeature.self).eventWriteContext { context, writer in
                 let builder = LogEventBuilder(
                     service: log.service ?? context.service,
                     loggerName: log.logger,
@@ -191,6 +191,7 @@ internal struct CrashLogReceiver: FeatureMessageReceiver {
 
     /// Time provider.
     let dateProvider: DateProvider
+    let logEventMapper: LogEventMapper?
 
     /// Process messages receives from the bus.
     ///
@@ -237,7 +238,7 @@ internal struct CrashLogReceiver: FeatureMessageReceiver {
 
         // crash reporting is considering the user consent from previous session, if an event reached
         // the message bus it means that consent was granted and we can safely bypass current consent.
-        core.scope(for: LogsFeature.name)?.eventWriteContext(bypassConsent: true) { context, writer in
+        core.scope(for: LogsFeature.self).eventWriteContext(bypassConsent: true) { context, writer in
             let event = LogEvent(
                 date: date,
                 status: .emergency,
@@ -285,7 +286,7 @@ internal struct CrashLogReceiver: FeatureMessageReceiver {
                 tags: nil
             )
 
-            writer.write(value: event)
+            logEventMapper?.map(event: event, callback: writer.write) ?? writer.write(value: event)
         }
 
         return true
@@ -309,7 +310,7 @@ internal struct WebViewLogReceiver: FeatureMessageReceiver {
         let tagsKey = LogEventEncoder.StaticCodingKeys.tags.rawValue
         let dateKey = LogEventEncoder.StaticCodingKeys.date.rawValue
 
-        core.scope(for: LogsFeature.name)?.eventWriteContext { context, writer in
+        core.scope(for: LogsFeature.self).eventWriteContext { context, writer in
             let ddTags = "\(versionKey):\(context.version),\(envKey):\(context.env)"
 
             if let tags = event[tagsKey] as? String, !tags.isEmpty {