+++ DatadogRUM/Tests/RUMMonitor/MonitorTests.swift
@@ -10,17 +10,7 @@ import DatadogInternal
 @testable import DatadogRUM
 
 class MonitorTests: XCTestCase {
-    private var core: PassthroughCoreMock! // swiftlint:disable:this implicitly_unwrapped_optional
-
-    override func setUp() {
-        super.setUp()
-        core = PassthroughCoreMock()
-    }
-
-    override func tearDown() {
-        core = nil
-        super.tearDown()
-    }
+    private let featureScope = FeatureScopeMock()
 
     func testWhenSessionIsSampled_itSetsRUMContextInCore() throws {
         // Given
@@ -28,8 +18,7 @@ class MonitorTests: XCTestCase {
 
         // When
         let monitor = Monitor(
-            core: core,
-            dependencies: .mockWith(core: core, sessionSampler: sampler),
+            dependencies: .mockWith(featureScope: featureScope, sessionSampler: sampler),
             dateProvider: DateProviderMock()
         )
         monitor.startView(key: "foo")
@@ -37,7 +26,7 @@ class MonitorTests: XCTestCase {
 
         // Then
         let expectedContext = monitor.currentRUMContext
-        let rumBaggage = try XCTUnwrap(core.context.baggages[RUMFeature.name])
+        let rumBaggage = try XCTUnwrap(featureScope.contextMock.baggages[RUMFeature.name])
         let rumContext = try rumBaggage.decode(type: RUMCoreContext.self)
         XCTAssertEqual(rumContext.applicationID, expectedContext.rumApplicationID)
         XCTAssertEqual(rumContext.sessionID, expectedContext.sessionID.toRUMDataFormat)
@@ -50,15 +39,14 @@ class MonitorTests: XCTestCase {
 
         // When
         let monitor = Monitor(
-            core: core,
-            dependencies: .mockWith(core: core, sessionSampler: sampler),
+            dependencies: .mockWith(featureScope: featureScope, sessionSampler: sampler),
             dateProvider: DateProviderMock()
         )
         monitor.startView(key: "foo")
         monitor.flush()
 
         // Then
-        XCTAssertNil(core.context.baggages[RUMFeature.name])
+        XCTAssertNil(featureScope.contextMock.baggages[RUMFeature.name])
     }
 
     func testStartView_withViewController_itUsesClassNameAsViewName() throws {
@@ -67,8 +55,7 @@ class MonitorTests: XCTestCase {
 
         // When
         let monitor = Monitor(
-            core: core,
-            dependencies: .mockWith(core: core, sessionSampler: .mockKeepAll()),
+            dependencies: .mockWith(featureScope: featureScope, sessionSampler: .mockKeepAll()),
             dateProvider: DateProviderMock()
         )
         monitor.startView(viewController: vc)
@@ -85,8 +72,7 @@ class MonitorTests: XCTestCase {
 
         // When
         let monitor = Monitor(
-            core: core,
-            dependencies: .mockWith(core: core, sessionSampler: .mockKeepAll()),
+            dependencies: .mockWith(featureScope: featureScope, sessionSampler: .mockKeepAll()),
             dateProvider: DateProviderMock()
         )
         monitor.startView(viewController: vc, name: "Some View")