+++ IntegrationTests/IntegrationScenarios/Scenarios/CrashReporting/CrashReportingWithRUMScenarioTests.swift
@@ -96,5 +96,8 @@ class CrashReportingWithRUMScenarioTests: IntegrationTests, RUMCommonAsserts {
         XCTAssertNotNil(try? lastRUMErrorEventJSON.value(forKeyPath: "error.meta") as [String: Any], "The error should include crash meta info")
         XCTAssertNotNil(try? lastRUMErrorEventJSON.value(forKeyPath: "error.threads") as [Any], "The error should include threads info")
         XCTAssertNotNil(try? lastRUMErrorEventJSON.value(forKeyPath: "error.binary_images") as [Any], "The error should include binary images info")
+
+        // Assert mapped value
+        XCTAssertEqual(crashRUMError.error.fingerprint, "mapped fingerprint", "Error should have a mapper supplied fingerprint value")
     }
 }